fixed changes that were missing from cherrypicked commit
