+++ uwrt_mars_rover_control/src/uwrt_mars_rover_control_node.cpp
@@ -87,7 +87,7 @@ int main(int argc, char** argv) {
 
   const bool default_realtime_mode = true;
   const bool use_realtime_kernel{
-      uwrt_mars_rover_utils::getParam(nh, "use_realtime_kernel", default_realtime_mode, node_name)};
+      uwrt_mars_rover_utils::getParam(nh, node_name, "use_realtime_kernel", default_realtime_mode)};
 
   if (use_realtime_kernel) {
     constexpr auto REALTIME_SCHEDULING_POLICY = SCHED_FIFO;
